# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FitnessTrack Public API
  version: 1.0.0
  description: >-
    REST API for the FitnessTrack platform. All endpoints respond with JSON and
    leverage RFC 7807 problem-details for errors.
servers:
  - url: /api/v1
    description: Version 1 base path
security:
  - bearerAuth: []
...
paths:
...
  /cycles:
    get:
      tags: [Cycles]
      operationId: listCycles
      summary: List cycles
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - name: subject_id
          in: query
          schema:
            type: integer
        - name: routine_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Cycle page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CyclePage'
    post:
      tags: [Cycles]
      operationId: createCycle
      summary: Create cycle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cycle'
      responses:
        '201':
          description: Created cycle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cycle'
  /cycles/{cycleId}:
    parameters:
      - name: cycleId
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Cycles]
      operationId: getCycle
      summary: Retrieve cycle
      responses:
        '200':
          description: Cycle document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cycle'
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Cycles]
      operationId: updateCycle
      summary: Update cycle
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CycleUpdate'
      responses:
        '200':
          description: Updated cycle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cycle'
    delete:
      tags: [Cycles]
      operationId: deleteCycle
      summary: Delete cycle
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
  /subject-routines:
    get:
      tags: [SubjectRoutines]
      operationId: listSubjectRoutines
      summary: List subject-routine links
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Link page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectRoutinePage'
    post:
      tags: [SubjectRoutines]
      operationId: createSubjectRoutine
      summary: Create link
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectRoutineCreate'
      responses:
        '201':
          description: Created link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectRoutine'
  /subject-routines/{linkId}:
    parameters:
      - name: linkId
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [SubjectRoutines]
      operationId: getSubjectRoutine
      summary: Retrieve link
      responses:
        '200':
          description: Link document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectRoutine'
        '404':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [SubjectRoutines]
      operationId: deleteSubjectRoutine
      summary: Delete link
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
      description: Page number (1-indexed)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
      description: Page size
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Comma separated sort fields, prefix with '-' for descending.
...
  responses:
    Problem:
      description: Problem details error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        errors:
          oneOf:
            - type: object
            - type: array
              items:
                type: object
    PageEnvelope:
      type: object
      properties:
        items:
          type: array
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required: [items, page, limit, total]
...
    Cycle:
      type: object
      properties:
        id:
          type: integer
        subject_id:
          type: integer
        routine_id:
          type: integer
        cycle_number:
          type: integer
        started_on:
          type: string
          format: date
          nullable: true
        ended_on:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CycleUpdate:
      type: object
      properties:
        started_on:
          type: string
          format: date
        ended_on:
          type: string
          format: date
        notes:
          type: string
    CyclePage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Cycle'
    SubjectRoutine:
      type: object
      properties:
        id:
          type: integer
        subject_id:
          type: integer
        routine_id:
          type: integer
        is_active:
          type: boolean
        saved_on:
          type: string
          format: date-time
    SubjectRoutineCreate:
      allOf:
        - $ref: '#/components/schemas/SubjectRoutine'
    SubjectRoutinePage:
      allOf:
        - $ref: '#/components/schemas/PageEnvelope'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SubjectRoutine'
...